lean/requirements.txt
django-compressor 
django
django-mpesa
djangorestframework